refactor(material_templates): move ensure_base_color_texture function for reusability
Was originally inside udim_to_mat_template, but the same function make sense to use for material template creation as well, since "white_default.dds" is the "base" diffuse texture for any vehicle materials in FS25
